Flickers + Meteors + Witchwood + Localization

+++ src/main/java/am2/api/event/SpellCastEvent.java
@@ -7,10 +7,10 @@ import net.minecraftforge.fml.common.eventhandler.Event;
 public class SpellCastEvent extends Event {
 	
 	public ItemStack spell;
-	public int manaCost;
+	public float manaCost;
 	public EntityLivingBase entityLiving;
 	
-	public SpellCastEvent(EntityLivingBase caster, ItemStack spell, int manaCost) {
+	public SpellCastEvent(EntityLivingBase caster, ItemStack spell, float manaCost) {
 		this.spell = spell;
 		this.manaCost = manaCost;
 		this.entityLiving = caster;
@@ -18,7 +18,7 @@ public class SpellCastEvent extends Event {
 	
 	public static class Pre extends SpellCastEvent {
 
-		public Pre(EntityLivingBase caster, ItemStack spell, int manaCost) {
+		public Pre(EntityLivingBase caster, ItemStack spell, float manaCost) {
 			super(caster, spell, manaCost);
 		}
 		
@@ -26,7 +26,7 @@ public class SpellCastEvent extends Event {
 	
 	public static class Post extends SpellCastEvent {
 
-		public Post(EntityLivingBase caster, ItemStack spell, int manaCost) {
+		public Post(EntityLivingBase caster, ItemStack spell, float manaCost) {
 			super(caster, spell, manaCost);
 		}
