+++ src/main/java/am2/items/SpellBase.java
@@ -62,7 +62,7 @@ public class SpellBase extends ItemSpellBase{
 	public void addInformation(ItemStack stack, EntityPlayer player, List<String> list, boolean par4){
 		if (!stack.hasTagCompound()) return;
 		
-		int manaCost = SpellUtils.getManaCost(stack);
+		float manaCost = SpellUtils.getManaCost(stack);
 		manaCost *= 1F + (float)((float)EntityExtension.For(player).getCurrentBurnout() / (float)EntityExtension.For(player).getMaxBurnout());
 		list.add("Mana Cost : " + manaCost);
 	}
